requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
fonttools==4.61.1
gunicorn==25.1.0
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
nltk==3.9.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
regex==2026.1.15
Sastrawi==1.0.1
scikit-learn==1.6.1
scipy==1.17.0
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tqdm==4.67.3
tzdata==2025.3
Werkzeug==3.1.5
xgboost==3.2.0
